refactor(modules): split discovery.rs into a directory by concern
discovery.rs was 2037 lines (1145 production, 892 tests) covering six
unrelated jobs. It is now a directory whose mod.rs holds the public API and
the scan itself, with the layers underneath split out:

  catalog.rs      the USB VID/PID table we recognize
  sysfs.rs        walking sysfs for tty/ALSA/net devices
  policy.rs       blocklist, quarantine, the bounded abandonable worker
  probe.rs        finding the interface that answers AT
  sim.rs          SIM identity, and the one optional CFUN repair
  claims.rs       which modems another subsystem already owns
  test_support.rs the two fake serial transports the tests share

Beyond relocation:

- find_alsa_card and find_net_iface were the same sysfs walk written twice;
  they now share find_in_iface_subdir. Their interface-name markers really
  do differ (":1." vs ":"), so that stays a parameter rather than being
  unified away — two new tests pin that difference down.
- read_lines_file_excerpt/read_volte_manifest_excerpt were line-for-line
  clones, as were active_vowifi_card_ids/active_volte_card_ids. The readers
  now share one generic read_manifest_excerpt, and the VoWiFi one resolves
  its path through lines_file_path() instead of repeating the env lookup.
- MockStream and ScriptedModem were both defined in the one test block;
  they are now in test_support.rs once.
- probe_is_at_capable and probe_sim_status were pub only so the module-wide
  test block could reach them. That block now sits beside each of them, so
  both are private; nothing outside this module ever called them.

DiscoveryPolicy's counters stay private: probe.rs asserts on the AT streak
through a #[cfg(test)] accessor rather than widening the field.

The only_the_one_shot_discover_scan_opts_into_sim_recovery guard reads its
own source with include_str! to check every scan_all_inner call site passes
SimRecovery::Disabled. Both its paths are updated for the new depth, and it
stays in mod.rs because that is where every one of those call sites lives.

Every one of the 42 original tests is preserved verbatim; the suite goes
from 1268 to 1270 solely because of the two new sysfs tests above.
